requirements.txt
python-dotenv
gradio
openai
pygame
requests
numpy
sounddevice
soundfile